MongoQuery extends Iterable, so all of related operations are supported now
